Instead of copy the class just for change the colour, should be better just modify the different part.
